+++ CHANGES.rst
@@ -5,7 +5,9 @@
 4.1.1 (unreleased)
 ==================
 
-- Nothing changed yet.
+- Expose the ``__doc__`` (and, where applicable, ``__bases__`` and
+  ``__dict__``) of the hooked object. This lets Sphinx document them.
+  See `issue 6 <https://github.com/zopefoundation/zope.hookable/issues/6>`_.
 
 
 4.1.0 (2017-07-26)